refactor(test): update test files
